backend/src/git_repo_documentor/agents/planner/file_identification.py
from google.adk.agents import LlmAgent
from google.adk.models import Gemini # Or your chosen LLM

# Placeholder for the actual tool - replace with real import later
# from ...tools.file_system import read_directory_tool
from google.adk.tools import FunctionTool

# --- Placeholder Tool ---
def placeholder_read_directory(path: str, recursive: bool = False) -> list[str]:
    """Placeholder function for reading directory."""
    print(f"Placeholder: Reading directory {path} (recursive={recursive})")
    # Return dummy data that matches expected type
    return [f"{path}/file1.py", f"{path}/subdir/file2.js", f"{path}/README.md"]

read_directory_tool = FunctionTool(
    func=placeholder_read_directory,
    description="Reads a directory and lists files/subdirectories."
)
# --- End Placeholder Tool ---


# Define the LLM model (as per plan)
# Consider making this configurable or passed in
common_model = Gemini(model="gemini-1.5-flash-latest") # Use a faster model initially if desired

# Define the Agent (as per plan)
file_identification_agent = LlmAgent(
    name="FileIdentificationAgent", # Added name for clarity
    model=common_model,
    instruction="Analyze the repository structure provided from 'repo_path' state. Identify primary code files suitable for documentation. Filter out test files, configuration files, and generated code unless explicitly requested. Focus on common source code file extensions.",
    tools=[read_directory_tool],
    output_key="identified_files", # Key where the list of files will be stored in the state
    # Add example or specific output format guidance if needed
    output_format="json", # Explicitly request JSON list if possible
    # Example of expected output for the LLM:
    # { "identified_files": ["src/main.py", "src/utils/helpers.py", "lib/auth.js"] }
)

print("FileIdentificationAgent defined.")

# Example usage (for testing purposes, usually run via Orchestrator/Runner)
async def _test_run():
    from google.adk.sessions import State
    initial_state = State({"repo_path": "/path/to/your/repo"})
    # Need ArtifactService for agent execution, even if dummy
    from google.adk.artifacts import InMemoryArtifactService
    artifact_service = InMemoryArtifactService()
    result_state = await file_identification_agent.run(initial_state, artifact_service)
    print("Test run result state:", result_state)
    # Expected output might look like:
    # Test run result state: {'repo_path': '/path/to/your/repo', 'identified_files': ['/path/to/your/repo/file1.py', '/path/to/your/repo/subdir/file2.js']} (depending on placeholder/LLM)

if __name__ == "__main__":
    import asyncio
    # asyncio.run(_test_run()) # Commented out as direct execution might need more setup
    pass
