backend/src/git_repo_documentor/agents/processing/testing_analyzer.py
from google.adk.agents import LlmAgent
from google.adk.models import Gemini # Or your chosen LLM
from google.adk.tools import FunctionTool
import os

# --- Placeholder Tools (Replace with actual imports) ---
def placeholder_read_directory(path: str, recursive: bool = False) -> list[str]:
    """Placeholder function for reading directory."""
    print(f"Placeholder: Reading directory {path} (recursive={recursive}) for tests")
    # Simulate finding test files related to the current file path
    dirname = os.path.dirname(path)
    filename = os.path.basename(path)
    base, ext = os.path.splitext(filename)
    # Common test locations/naming conventions
    potential_tests = [
        os.path.join(dirname, f"test_{base}{ext}"),
        os.path.join(os.path.dirname(dirname), "tests", f"test_{base}{ext}"),
        os.path.join(dirname, "tests", f"test_{filename}"),
        os.path.join(os.path.dirname(dirname), "test", f"Test{base.capitalize()}.java"), # Java example
    ]
    # Return plausible test file paths for the demo
    return [p for p in potential_tests if "utils" in p or "app" in p] # Filter for demo

read_directory_tool = FunctionTool(
    func=placeholder_read_directory,
    description="Reads a directory to find potential test files related to a source file."
)

def placeholder_read_file(path: str) -> str:
    """Placeholder function for reading file content."""
    print(f"Placeholder: Reading test file {path}")
    if "test_utils.py" in path:
        return """
import pytest
from ..src import utils

def test_format_name():
    assert utils.format_name({"name": "Test"}) == "User: Test"

def test_clean_text():
    assert utils.clean_text("  text  ") == "text"

# Test coverage seems decent based on this snippet
"""
    elif "test_app.py" in path:
        return """
import app

def test_main_runs(mocker):
    mocker.patch('app.run_server')
    app.main()
    app.run_server.assert_called_once()
"""
    else:
        return "Test content not found for placeholder."

read_file_tool = FunctionTool(
    func=placeholder_read_file,
    description="Reads the content of a specific test file."
)
# --- End Placeholder Tools ---

# Define the LLM model
common_model = Gemini(model="gemini-1.5-flash-latest")

# Define the Agent
testing_analyzer_agent = LlmAgent(
    name="TestingAnalyzerAgent",
    model=common_model,
    instruction="""Identify and analyze test files related to the source file specified in 'current_file_path'.
1. Use 'placeholder_read_directory' to search for potential test files in common locations (e.g., sibling 'test_*.py', '../tests/', etc.). Look for files matching patterns like 'test_*', '*_test', '*Test.java'.
2. If test files are found, use 'placeholder_read_file' to read their content.
3. Analyze the test file content to summarize the testing approach (e.g., framework used like pytest, JUnit), estimate test coverage (e.g., high, medium, low based on number/nature of tests), and identify key test cases or areas tested.
4. If no test files are found, state that.
Output the analysis as a concise summary string.""",
    tools=[read_directory_tool, read_file_tool],
    output_key="testing_analysis", # Key for the testing summary
    output_format="text", # Simple summary string
    # Example input state:
    # { "current_file_path": "/path/to/repo/src/utils.py", ... }
    # Example expected output:
    # { ..., "testing_analysis": "Test file found at ../tests/test_utils.py. Uses pytest framework. Appears to have good coverage for format_name and clean_text functions." }
    # OR
    # { ..., "testing_analysis": "No specific test files found for src/utils.py." }
)

print("TestingAnalyzerAgent defined.")

# Example usage (for testing purposes)
async def _test_run():
    from google.adk.sessions import State
    from google.adk.artifacts import InMemoryArtifactService

    initial_state = State({
        "current_file_path": "/path/to/repo/src/utils.py",
        # Context from previous steps might exist here
    })
    artifact_service = InMemoryArtifactService()
    result_state = await testing_analyzer_agent.run(initial_state, artifact_service)
    print("Test run result state (utils):", result_state)

    initial_state_no_test = State({
        "current_file_path": "/path/to/repo/src/config_loader.py",
    })
    result_state_no_test = await testing_analyzer_agent.run(initial_state_no_test, artifact_service)
    print("Test run result state (no test):", result_state_no_test)


if __name__ == "__main__":
    import asyncio
    # asyncio.run(_test_run())
    pass
